feat: Add Skapi, Banking, and DesignSystem components
Add new components Skapi, Banking, and DesignSystem to the project. These components render custom content for each respective project. This enhances the modularity and reusability of the codebase.

Refactor CarouselSlide component to include a link to the project page for each slide. This allows users to navigate to the individual project pages directly from the carousel.

Update Carousel component viewport padding to improve responsiveness. The new padding ensures that the carousel is properly displayed on different screen sizes.

Update ContactForm layout and styles for improved readability and responsiveness. This enhances the user experience when filling out the contact form.

Refactor WorkSection component to improve responsiveness and initial height calculation. This ensures that the component adapts well to different screen sizes and provides a better user experience.

Update Footer background color to use slate-950 instead of black. This change aligns the footer with the overall color scheme of the project.

Fix minor CSS issue in WelcomeLinks component. The \Resume\ link has been corrected to display \Contact\ instead.

Update npm dependencies to the latest stable versions.