refactor: Update background color in various components for consistency
Update the background color in the Footer, ClippedGradientTicker, and ContactForm components to use the new color \bg-stone-950\ instead of \bg-black\. This change ensures consistency with the overall color scheme of the project and improves the visual appeal of these components.

Improve responsiveness and initial height calculation in the WorkSection component. This update ensures that the component adapts well to different screen sizes and provides a better user experience.

Update the WelcomeLinks component to display \Contact\ instead of \Resume\ for the corresponding link. This fix corrects a minor CSS issue and improves the accuracy of the displayed link.

Refactor the Carousel component to include project page links for each slide. This enhancement allows users to navigate directly to the individual project pages from the carousel.

Update npm dependencies to the latest stable versions for improved performance and security.