feat: Add ChallengeSection component for displaying challenges in ProjectPage
This commit adds a new component called ChallengeSection to the ProjectPage. The ChallengeSection component is responsible for rendering a section that displays the challenges of a project. It takes an overview string and an array of Challenge objects as props.

The ChallengeSection component includes a function called pairChallenges, which pairs the challenges into groups of two. It then renders the challenges in a responsive layout using the ChallengeCard component.

This addition enhances the ProjectPage by providing a visually appealing and organized way to showcase the challenges of a project.